refactor: UserService 추상화
